Title: 비율 구하기, Time: 700ms, Memory: 23MB, Status: Passed - CodeTree

+++ 240613/비율 구하기/find-proportions.java
@@ -13,7 +13,7 @@ public class Main {
         }
 
         for(Map.Entry<String, Integer> e : m.entrySet()) {
-            double ratio = e.getValue() * 100 / n;
+            double ratio = (double) e.getValue() * 100 / n;
             System.out.printf("%s %.4f\n", e.getKey(), ratio);
         }
     }
